+++ src/Main/handlers/PlayerPing.h
@@ -45,7 +45,7 @@ namespace Game
                     CMessage updatePingAck(player_session->GetEncryptionKey());
                     updatePingAck.SetSession(id);
                     updatePingAck.SetCommand(callback.message->GetOrder() + 1, callback.message->GetMission() + 1, callback.message->GetExtra(), callback.message->GetOption());
-                    updatePingAck.SetData(reinterpret_cast<uint8_t*>(update_data.data()), update_data.size());
+                    updatePingAck.SetData(reinterpret_cast<uint8_t*>(update_data.data()), static_cast<std::uint16_t>(update_data.size()));
                     player_session->Send(updatePingAck);
                 }
             };
